+++ client/src/components/CreateCharacter.jsx
@@ -8,6 +8,8 @@ import CharacterCard from './CharacterCard';
 import { API } from '../config/api';
 import { getCampaign, updateCampaignCharacter } from '../actions/campaignActions';
 import { useParams } from 'react-router-dom';
+import SelectedCharacterImage from './SelectedCharacterImage';
+import SelectedCharacterColor from './SelectedCharacterColor';
 
 const CreateCharacter = ({ character }) => {
     const [name, setName] = useState(character.name);
@@ -129,12 +131,8 @@ const CreateCharacter = ({ character }) => {
                         />
                     </Box>
                     <Box display='flex' gap={2}>
-                        <Box height='100px' width='100px' backgroundColor='orangered'>
-                            Select character image (TODO)
-                        </Box>
-                        <Box height='100px' width='200px' backgroundColor='yellowgreen'>
-                            Select color (TODO)
-                        </Box>
+                        <SelectedCharacterImage currentImage={chosenImage} onSelectImage={image => setChosenImage(image)} />
+                        <SelectedCharacterColor currentColor={chosenColor} onSelectColor={color => setChosenColor(color)} />
                     </Box>
                 </Box>
             </Paper>